requirements.txt
fastapi
uvicorn
jinja2
python-multipart
websockets
opencv-python
ollama
aiofiles
sentence-transformers
faiss-cpu
tqdm

